+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New option to disable sending read receipts for messages (anonymous message read).
 - Added like and pass buttons to Cupid's Picks profiles.
 - New option to add a high resolution photo upload button on the site.
-- Ability to view profile who viewed you (needs to use manually to update the list)
+- Button to fetch all interested profiles (including likes you) and view them in the interested profiles page.
 
 ### Fixed
 - likes remaining count not updating correctly.